+++ src/solver_options.jl
@@ -57,7 +57,7 @@ mutable struct OptimizerOptions
     presolve_bt_algo::Any                                       # Method used for bound tightening procedures, can either be an index of default methods or functional inputs
     presolve_bt_relax_integrality::Bool                         # Relax the MIP solved in built-in relaxation scheme for time performance
     presolve_bt_mip_time_limit::Float64                         # Time limit for a single MIP solved in the built-in bound tightening algorithm (with partitions)
-    use_start_as_local_solution::Bool                           # Use starting value as local solution for presolve without invoking a local solver
+    use_start_as_incumbent::Bool                           # Use starting value as a local incumbent solution for presolve without invoking a local solver
 
     # Domain Reduction
     presolve_bp::Bool                                           # Conduct basic bound propagation
@@ -109,7 +109,7 @@ function get_default_options()
     presolve_bt_algo = 1
     presolve_bt_relax_integrality = false
     presolve_bt_mip_time_limit = Inf
-    use_start_as_local_solution = false
+    use_start_as_incumbent = false
     presolve_bp = false
 
     return OptimizerOptions(
@@ -153,7 +153,7 @@ function get_default_options()
         presolve_bt_algo,
         presolve_bt_relax_integrality,
         presolve_bt_mip_time_limit,
-        use_start_as_local_solution,
+        use_start_as_incumbent,
         presolve_bp,
     )
 end